fix: split named-factory layout helpers to satisfy backend-size gate
Move MapNamedFactoryDir, ListNamedFactories, DeleteNamedFactory, and related helpers out of layout_expansion.go so required CI backend-size passes.
